refactor(icon_button): `customLayout` Modifier 확장 함수 제거
- `PickleIconButton`에서 사용하던 `customLayout` Modifier 확장 함수를 제거합니다.
- Modifier 확장 함수 대신 `Modifier.layout`을 직접 사용하여 로직을 `PickleIconButton` 내부로 이동시킵니다.
